+++ busy_beaver/blueprints/__init__.py
@@ -1,4 +1,4 @@
 from .healthcheck import healthcheck_bp  # noqa
 from .integration import integration_bp  # noqa
+from .poller import poller_bp  # noqa
 from .slack import slack_bp  # noqa
-from .tasks import tasks_bp  # noqa
